Info about project!
## Research Question:

Using a rich dataset of survey data of public libraries across the US, we decided to investigate the question: "How does state funding impact the number of libraries/staff/resources available per state?"

Public libraries are funded from a variety of sources across the nation, based on factors ranging from the structure of local government to the physical infrastructure needs of library systems. A number of different metrics can be used to measure the scope of library service: the population of the area served, the volume of system circulation, and many others. Lastly, we believe information on operating expenditures (staff and circulation expenditures) can be used to estimate service efficiency in conjunction with other data.

## We are taking into account:

- number of libaries/branches of libraries/bookmobiles for count
- funding from the government
- population per state
- how many libarary staff & their salaries

## The Data:

Our dataset comes from the 2018 Public Libraries Survey (PLS), administered by the Institute of Museum and Library Services. This large and exceptionally well-curated dataset is accompanied by a robust data dictionary, and contains "information about about library visits, circulation, ... staffing, electronic resources, operating revenues and expenditures..." for nearly every public library system in the United States and territories.

We chose to focus on a subset of the data which collated primarily quantitative data by state. (Far more detailed data on each of the nation's >9,000 library systems, and >13,000 branch locations is also available in the data set; closely related datasets are available from 2018 back to 1992.)

Of note, the data has already been extensively cleaned; in particular, a series of imputation codes has been used to denote which of various methods were used to address missing and null values. A large part of the cleaning we did was to remove these imputation codes, since this information was not critical to our analysis.

## Source: 

https://www.imls.gov/research-evaluation/data-collection/public-libraries-survey